refactor: 당첨금 enum을 모델 패키지로 이동하여 구조 개선
