+++ src/main/java/lotto/custom/service/CalculateYieldService.java
@@ -3,7 +3,7 @@ package lotto.custom.service;
 import static lotto.custom.constants.NumberConstants.LOTTO_PRICE;
 
 import java.util.List;
-import lotto.custom.constants.NumberConstants;
+import lotto.custom.model.LottoPrize;
 import lotto.custom.model.Lottos;
 
 public class CalculateYieldService {
@@ -12,7 +12,7 @@ public class CalculateYieldService {
         int index = 0;
         double sum = 0;
 
-        for (NumberConstants.Prize prize : NumberConstants.Prize.values()) {
+        for (LottoPrize prize : LottoPrize.values()) {
             sum += prize.getPrizeMoney() * result.get(index++);
         }
